feat(ui): render app with miro provider
