+++ web/client/app.html
@@ -16,7 +16,7 @@
     <div id="root"></div>
 
     <script
-      src="src/app/App.tsx"
+      src="src/main.tsx"
       type="module"></script>
   </body>
 </html>
